Makes Jaiqu use Burr
[Burr](https://github.com/dagworks-inc/burr) will enable someone to more easily debug
the state of the jaiqu application if it's giving nonsense back
or it hits the maxium number of retries.

The design decision here is to make a close port of the original
code. It could be broken down further, but that would require
more work and I wasn't sure it TQDM was something that
had to stay or not.

This is a 100% backwards compatible change with the
current API.

tests have been moved to live under /tests/ and match
the README.

The mermaid image has been replaced with the
actual application structure that Burr generates.
