+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "jaiqu"
-version = "0.0.4"
+version = "0.0.5"
 authors = [
   { name = "Alex Reibman", email = "areibman@gmail.com" },
   { name = "Howard Gil", email = "howardbgil@gmail.com" },
@@ -21,7 +21,8 @@ classifiers = [
 dependencies = [
     "jq==1.6.0",
     "openai~=1.12.0",
-    "jsonschema==4.21.1"
+    "jsonschema==4.21.1",
+    "burr~=0.*",
 ]
 
 [project.optional-dependencies]
@@ -29,6 +30,7 @@ dev = [
     "pytest>=7.4.4",
     "flake8>=3.1.0",
     "coverage[toml]>=7.4.0",
+    "burr[start]~=0.*",
 ]
 
 [project.urls]
